+++ src/vscode-textarea.ts
@@ -6,10 +6,7 @@ import {styleMap} from 'lit/directives/style-map.js';
 import {VscElement} from './includes/VscElement';
 
 /**
- * An example element.
- *
- * @slot - This element has a slot
- * @csspart button - The button
+ * Multi-line text input.
  */
 @customElement('vscode-textarea')
 export class VscodeTextarea extends VscElement {
@@ -52,6 +49,10 @@ export class VscodeTextarea extends VscElement {
   @property()
   spellcheck = false;
 
+  /**
+   * Use monospace fonts. The font family, weight, size, and color will be the same as set in the
+   * VSCode code editor.
+   */
   @property({type: Boolean, reflect: true})
   monospace = false;
 
@@ -64,6 +65,9 @@ export class VscodeTextarea extends VscElement {
     return this._value;
   }
 
+  /**
+   * Getter for the inner textarea element if it needs to be accessed for some reason.
+   */
   get wrappedElement(): HTMLTextAreaElement {
     return this._textareaEl;
   }
@@ -77,6 +81,12 @@ export class VscodeTextarea extends VscElement {
 
   @state()
   private _value = '';
+
+  @state()
+  private _textareaPointerCursor = false;
+
+  @state()
+  private _shadow = false;
 
   private _handleChange(ev: InputEvent) {
     this._value = this._textareaEl.value;
@@ -98,13 +108,56 @@ export class VscodeTextarea extends VscElement {
     );
   }
 
+  private _handleMouseMove(ev: MouseEvent) {
+    if (this._textareaEl.clientHeight >= this._textareaEl.scrollHeight) {
+      this._textareaPointerCursor = false;
+      return;
+    }
+
+    const SCROLLBAR_WIDTH = 14;
+    const BORDER_WIDTH = 1;
+    const br = this._textareaEl.getBoundingClientRect();
+    const x = ev.clientX;
+
+    this._textareaPointerCursor =
+      x >= br.left + br.width - SCROLLBAR_WIDTH - BORDER_WIDTH * 2;
+  }
+
+  private _handleScroll() {
+    this._shadow = this._textareaEl.scrollTop > 0;
+  }
+
   static get styles(): CSSResultGroup {
     return [
       super.styles,
       css`
         :host {
-          display: block;
+          display: inline-block;
+          height: 40px;
+          position: relative;
           width: 320px;
+        }
+
+        :host([cols]) {
+          width: auto;
+        }
+
+        :host([rows]) {
+          height: auto;
+        }
+
+        .shadow {
+          box-shadow: var(--vscode-scrollbar-shadow) 0 6px 6px -6px inset;
+          display: none;
+          inset: 0 0 auto 0;
+          height: 6px;
+          pointer-events: none;
+          position: absolute;
+          width: 100%;
+        }
+
+        .shadow.visible {
+          display: block;
         }
 
         textarea {
@@ -122,6 +175,16 @@ export class VscodeTextarea extends VscElement {
           );
           font-size: var(--vscode-font-size, 13px);
           font-weight: var(--vscode-font-weight, normal);
+          height: 100%;
+          width: 100%;
+        }
+
+        :host([cols]) textarea {
+          width: auto;
+        }
+
+        :host([rows]) textarea {
+          height: auto;
         }
 
         textarea.monospace {
@@ -129,6 +192,10 @@ export class VscodeTextarea extends VscElement {
           font-family: var(--vscode-editor-font-family);
           font-size: var(--vscode-editor-font-size);
           font-weight: var(--vscode-editor-font-weight);
+        }
+
+        textarea.cursor-pointer {
+          cursor: pointer;
         }
 
         textarea:focus {
@@ -141,7 +208,7 @@ export class VscodeTextarea extends VscElement {
         }
 
         textarea::-webkit-scrollbar {
-          width: 15px;
+          width: 14px;
         }
 
         textarea::-webkit-scrollbar-thumb {
@@ -161,12 +228,35 @@ export class VscodeTextarea extends VscElement {
             rgba(100, 100, 100, 0.7)
           );
         }
+
+        textarea::-webkit-scrollbar-thumb:active {
+          background-color: var(
+            --vscode-scrollbarSlider-activeBackground,
+            rgba(100, 100, 100, 0.7)
+          );
+        }
+
+        textarea::-webkit-scrollbar-corner {
+          background-color: transparent;
+        }
+
+        textarea::-webkit-resizer {
+          background-image: url('data:image/png;base64,iVBORw0KGgoAAAANSUhEUgAAAAcAAAAHCAYAAADEUlfTAAAAAXNSR0IB2cksfwAAAAlwSFlzAAALEwAACxMBAJqcGAAAACJJREFUeJxjYMAOZuIQZ5j5//9/rJJESczEKYGsG6cEXgAAsEEefMxkua4AAAAASUVORK5CYII=');
+          background-repeat: no-repeat;
+          background-position: right bottom;
+        }
       `,
     ];
   }
 
   render(): TemplateResult {
     return html`
+      <div
+        class=${classMap({
+          shadow: true,
+          visible: this._shadow,
+        })}
+      ></div>
       <textarea
         autocomplete=${this.autocomplete}
         ?disabled=${this.disabled}
@@ -174,6 +264,7 @@ export class VscodeTextarea extends VscElement {
         id="textarea"
         class=${classMap({
           monospace: this.monospace,
+          'cursor-pointer': this._textareaPointerCursor,
         })}
         maxlength=${ifDefined(this.maxlength)}
         minlength=${ifDefined(this.minlength)}
@@ -189,6 +280,8 @@ export class VscodeTextarea extends VscElement {
         ?spellcheck=${this.spellcheck}
         @change=${this._handleChange}
         @input=${this._handleInput}
+        @mousemove=${this._handleMouseMove}
+        @scroll=${this._handleScroll}
       >
 ${this._value}</textarea
       >